task 7

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -8,7 +8,7 @@
 int print_last_digit(int n)
 {
 if (n < 0)
-n = -((unsigned int)(n));
+n = -((unsigned long int)(n));
 n = n % 10;
 _putchar('0' + n);
 return (n);
